Avoid dependency on coincidental serialization
Prior to this commit, the `TransparentExponentialBackoffSupervisor` actor depended
on the fact that the `Stop` directive will always generate a message that comes after
any messages sent to self during the execution of the decider.

With this change we immediately update our state to handle the `Termination` rather than
assuming our message will change the state before the `Termination` message arrival.